Fix issue with buffer cutting off JSON
About 20% of the logs that are coming in exceed the buffer size that was
set. Also, I learnt that there is a better way to handle JSON payload
with Go.

Using the decoder means that implementing the Unmarshaller interface
allows the same behavior in a safer way.

The decoder also supports stream so it will parse the content no matter
the size of the JSON payload.